Remove traces of the :key option
Basing on code examples in the documentation and Attr Encrypted's
Readme, I believe that this option was intended to hold some encryption
key, possibly to be used with key derivation functions. Such feature
must be very useful in Attr Encrypted gem, but does not fit well in Attr
Masker.

Let's remove it completely.