Do not evaluate maskers prematurely
So far, all the objects passed as options to the ::attr_masker method
were eagerly evaluated, if they were responding to #call or were
symbols denoting some model methods.

Actually, that was the source of problems and could result in a
premature masker evaluation.  In the worst case, that could cause the
unexpected behaviour when checking the :if and :unless options, as the
option evaluation order was arbitrary.

This is fixed in this commit.